Introducing run configuration test helper Co-authored-by: Jonas Mende <jonas.mende@sky.uk>
