+++ controllers/pipelines/runconfiguration_controller_decoupled_test.go
@@ -488,38 +488,6 @@ var _ = Describe("RunConfiguration controller k8s integration", Serial, func() {
 	})
 })
 
-func updateOwnedSchedules(runConfiguration *pipelinesv1.RunConfiguration, updateFn func(schedule *pipelinesv1.RunSchedule)) error {
-	ownedSchedules, err := findOwnedRunSchedules(ctx, k8sClient, runConfiguration)
-	if err != nil {
-		return err
-	}
-
-	for _, ownedSchedule := range ownedSchedules {
-		updateFn(&ownedSchedule)
-		Expect(k8sClient.Status().Update(ctx, &ownedSchedule)).To(Succeed())
-	}
-
-	return nil
-}
-
-func matchRunConfiguration(runConfiguration *pipelinesv1.RunConfiguration, matcher func(Gomega, *pipelinesv1.RunConfiguration)) func(Gomega) {
-	return func(g Gomega) {
-		g.Expect(k8sClient.Get(ctx, runConfiguration.GetNamespacedName(), runConfiguration)).To(Succeed())
-		matcher(g, runConfiguration)
-	}
-}
-
-func matchSchedules(runConfiguration *pipelinesv1.RunConfiguration, matcher func(Gomega, *pipelinesv1.RunSchedule)) func(Gomega) {
-	return func(g Gomega) {
-		ownedSchedules, err := findOwnedRunSchedules(ctx, k8sClient, runConfiguration)
-		g.Expect(err).NotTo(HaveOccurred())
-		g.Expect(ownedSchedules).NotTo(BeEmpty())
-		for _, ownedSchedule := range ownedSchedules {
-			matcher(g, &ownedSchedule)
-		}
-	}
-}
-
 func matchRuns(runConfiguration *pipelinesv1.RunConfiguration, matcher func(Gomega, *pipelinesv1.Run)) func(Gomega) {
 	return func(g Gomega) {
 		ownedRuns, err := findOwnedRuns(ctx, k8sClient, runConfiguration)
@@ -545,54 +513,6 @@ func hasNoSchedules(runConfiguration *pipelinesv1.RunConfiguration) func(Gomega)
 		g.Expect(err).NotTo(HaveOccurred())
 		g.Expect(ownedSchedules).To(BeEmpty())
 	}
-}
-
-func createSucceededRcWithSchedule() *pipelinesv1.RunConfiguration {
-	runConfiguration := createStableRcWith(func(runConfiguration *pipelinesv1.RunConfiguration) *pipelinesv1.RunConfiguration {
-		runConfiguration.Spec.Triggers = pipelinesv1.RandomScheduleTrigger()
-		return runConfiguration
-	}, apis.Updating)
-
-	Eventually(matchSchedules(runConfiguration, func(g Gomega, ownedSchedule *pipelinesv1.RunSchedule) {
-		g.Expect(ownedSchedule.Status.SynchronizationState).To(Equal(apis.Creating))
-	})).Should(Succeed())
-
-	Expect(updateOwnedSchedules(runConfiguration, func(ownedSchedule *pipelinesv1.RunSchedule) {
-		ownedSchedule.Status.SynchronizationState = apis.Succeeded
-	})).To(Succeed())
-
-	Eventually(matchRunConfiguration(runConfiguration, func(g Gomega, fetchedRc *pipelinesv1.RunConfiguration) {
-		g.Expect(runConfiguration.Status.SynchronizationState).To(Equal(apis.Succeeded))
-	})).Should(Succeed())
-
-	return runConfiguration
-}
-
-func createSucceededRc() *pipelinesv1.RunConfiguration {
-	return createStableRcWith(func(runConfiguration *pipelinesv1.RunConfiguration) *pipelinesv1.RunConfiguration {
-		return runConfiguration
-	}, apis.Succeeded)
-}
-
-func createSucceededRcWith(modifyRc func(runConfiguration *pipelinesv1.RunConfiguration) *pipelinesv1.RunConfiguration) *pipelinesv1.RunConfiguration {
-	return createStableRcWith(modifyRc, apis.Succeeded)
-}
-
-func createStableRcWith(modifyRc func(runConfiguration *pipelinesv1.RunConfiguration) *pipelinesv1.RunConfiguration, synchronizationState apis.SynchronizationState) *pipelinesv1.RunConfiguration {
-	runConfiguration := pipelinesv1.RandomRunConfiguration()
-	runConfiguration.Spec.Run.RuntimeParameters = []pipelinesv1.RuntimeParameter{}
-	runConfiguration.Spec.Triggers = pipelinesv1.Triggers{}
-	modifiedRc := modifyRc(runConfiguration)
-	Expect(k8sClient.Create(ctx, modifiedRc)).To(Succeed())
-
-	Eventually(matchRunConfiguration(modifiedRc, func(g Gomega, fetchedRc *pipelinesv1.RunConfiguration) {
-		g.Expect(fetchedRc.Status.ObservedGeneration).To(Equal(modifiedRc.Generation))
-		g.Expect(fetchedRc.Status.SynchronizationState).To(Equal(synchronizationState))
-		g.Expect(fetchedRc.Status.Conditions.SynchronizationSucceeded().Reason).To(BeEquivalentTo(synchronizationState))
-		modifiedRc = fetchedRc
-	})).Should(Succeed())
-
-	return modifiedRc
 }
 
 func HavePipelineVersion(version string) types.GomegaMatcher {
